Remove translation container customization
The translation container customization allowed users to declare an
alternative field (the default was `:translations`) as the field that
contained the translations for the schema.

This customization has been sacrificed to ease the rewrite of the
QueryBuilder module. Now, Trans will always look in the `:translations`
field and there is no way to customize it.