Refactor QueryBuilder module
All the functions of the `Trans.QueryBuilder` module have been removed
and replaced with the macro `translated/2`. This macro generates
an SQL Fragment and can be directly used from inside an `Ecto.Query`.

This removes the necessity of building the query before calling the
`Trans.QueryBuilder` functions and integrates seamelessly with
`Ecto.Query` provided functions for comparisons, etc.
This also permits adding conditions on joined schemas.