Fix PostgreSQL JSONB operators precedence errors
In PostgreSQL 9.4 [1] the operators had different precedence rules than in
PostgreSQL 9.5 or higher [2] (take a look at "any other operator" in the table.

This differences caused errors when triggering a query that contained something
similar to `where a->b is not null`, which 9.4 would interpret as
`where a->(b is not null)` and 9.5 or higher would interpret as
`where (a->b) is not null`.

This commits adds parentheses around the JSONB accesses to avoid this kind of
problems by stating the desired preference explicitly.

[1] https://www.postgresql.org/docs/9.4/static/sql-syntax-lexical.html
[2] https://www.postgresql.org/docs/9.5/static/sql-syntax-lexical.html