Allow passing the locale from variables
Previously Trans was cally `to_string` in the locale before it was
unquoted. Since strins and atoms are represented as themselves in the
AST they were working fine. Trouble came when passed a variable, which
is represented in the AST as a three-element tuple and raised an error
when trying to convert it to a string.