Migrate to CircleCI
TravisCI started (but not yet completed) the migration from GitHub services to GitHub apps. In the meantime this means that we can not perform the migration manually nor see the the build status in the GitHub interface.

I am taking this as an opportunity to migrate to a new CI provider (CircleCI in this case). This will restore the ability to see build statuses in GitHub and will provide faster builds.

This is also a good moment to update the list of supported Elixir versions. The new minimum version is Elixir 1.4.
This aligns Trans with the requirements of the upcoming Phoenix 1.4 and Ecto 3.0.