Use Ecto 3.0
Ecto 3.0 splits itself into `ecto` and `ecto_sql` dependencies, which
can be run separately. Trans only depends on `ecto`. `ecto_sql` is an
optional dependency which will provide the `Trans.QueryBuilder` module.

Trans version has been bumped to 2.1 since the dependency update is a
breaking change. Despite of this, client code that is using Trans does not
need to be changed to use the new version (only the required modifications
for Ecto 3.0 are needed).