Return nil when a valid locale has no translations Per the previous commit, the Postgres jsonb type returns either a JSON 'null' or SQL NULL when there is not data. The cases are:
trans_test=# select translations->'it' from articles;
 ?column?
----------
 null
(1 row)

trans_test=# select translations->'au' from articles;
 ?column?
----------
 ∅
(1 row)

trans_test=# select translations->'fr'->>'thing' from articles;
 ?column?
----------
 ∅
(1 row)

trans_test=# select translations->'it'->>'title' from articles;
 ?column?
----------
 ∅
(1 row)

This PR coerces JSON `null` in the first example to be
a SQL NULL. Therefore all cases return `nil` consistently in the
Elixir context which is a more expected outcome.

All original tests still passed but that may be more
related to the shape of the test data.

This PR updates the test that check for a locale having
translations to use `is_nil/1` rather than `!= 'null'`.

It needs to be determined if this is a breaking change
or not.

Other than documentation updates, this PR completes
the work defined in the original proposal.