Implement dynamic fallback chains for translated/3 This commit revisits Trans.QueryBuilder with the following aims:
1. Support a list of locales (a fallback chain), not
   just a single local in the `translated/3` macro.
2. Support using `translated/3` as part of a query's
   `:select` clause such that the correct translation
	 or the original base value is returned
3. That such a translated column is returned in such
   a way that it will load correctly into an Ecto schema
	 by Ecto without further efforts.

The intent overall is to make for a simple and straight
forward way of retrieving translated columns that is little
different from a non-translated column. Consider the following
example using the Article schema used for Trans tests.

Firstly, a query as if the data was not translatable might be:

  from(
    a in Article,
  	select: [a.title, a.body]
  )

To turn this into a translated query we can say:

  from(
    a in Article,
    select: [translated(Article, a.title, [:de, :es]), translated(Article, a.body, [:de, :es]))]
  )

While this will correctly execute, the column names won't
match what Ecto expects when it loads the schemas. To make
that happen, the expression needs include a column alias which
would look like:

  fragment("? AS ?", translated(Article, a.title, [:de, :es]), :title)

To make this easier, a second macro, translated_as/3 is exactly the same
as `translated/3` but it adds the column alias automatically. Now
our translation query becomes:

  from(
    a in Article,
    select: [translated_as(Article, a.title, [:de, :es]), translated_as(Article, a.body, [:de, :es]))]
  )

and the data will load correctly into an Ecto schema without any further work.

The current released version of Trans indicates that all queries
should return the underlying base column if a translated value
does not exist.  This implementation honours that requirement,
if no translation exists in the fallback chain, the underlying
base column value is returned.

There is one exception to this rule that is also consistent
with the current implementation. If the query is on the translations
container itself, and no translations can be found for the given
locale(s) then NULL is returned.

In implementing these enhancements it was expected that the use
of the `dynamic` macro would allow runtime determination of the
code required to manage fallbacks. Unfortunately, `dynamic` cannot
be used:

1. `dynamic` cannot be used in the `:select` clause, meaning goal
   (2) cannot be achieved
2. `dynamic` must be used at the root of a clause (like `:where`).
   Which means `is_nil(^dynamic([a], query))` is not permitted

Therefore this implementation relies upon a database function
called `translate_field`. To install this function in the database
a migration is required. This commit also adds a mix task called
`trans_gen_translate_function` that generates the migration and
must be created before any translate queries with dynamic locales
can be used.

This does introduce a migration dependency for existing code however
the path is simple:

* mix deps.update trans
* mix trans_gen_translate_function
* mix ecto.migrate

This is a onetime action.  All the existing tests pass without
change and there is no change to existing application code
required.